% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.polly-package.R
\docType{package}
\name{aws.polly-package}
\alias{aws.polly}
\alias{aws.polly-package}
\title{aws.polly}
\description{
Client for AWS Polly
}
\details{
This is a client for AWS Polly, a speech synthesis service.
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\seealso{
\href{http://aws.amazon.com/documentation/polly}{AWS Polly Documentation}
}
\keyword{package}

