% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicons.R
\name{get_lexicon}
\alias{get_lexicon}
\alias{put_lexicon}
\alias{delete_lexicon}
\title{Lexicons}
\usage{
get_lexicon(lexicon, token, ...)

put_lexicon(lexicon, content, ...)

delete_lexicon(lexicon, ...)
}
\arguments{
\item{lexicon}{A character string specifying the name of a lexicon. If missing, a list of available lexicons is returned.}

\item{token}{Optionally, a pagination token.}

\item{\dots}{Additional arguments passed to \code{\link{pollyHTTP}}.}

\item{content}{A character string containing the content of the PLS lexicon.}
}
\value{
A list.
}
\description{
List, put, and delete lexicons
}
\details{
Note: \code{put_lexicon} will overwrite an existing lexicon with the same name.
}
\examples{
\dontrun{
list_lexicons()
}
}
