% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_voices.R
\name{list_voices}
\alias{list_voices}
\title{List available voices}
\usage{
list_voices(language = "en-US", token, ...)
}
\arguments{
\item{language}{An ISO 3166 country identification tag.}

\item{token}{Optionally, a pagination token.}

\item{\dots}{Additional arguments passed to \code{\link{pollyHTTP}}.}
}
\value{
A data frame of available names.
}
\description{
Retrieve a list of available voices
}
\examples{
\dontrun{
list_voices(language = "cy-GB")
}
}

